+++ src/auth/client/top-level-interaction.ts
@@ -1,9 +1,11 @@
 // Copied from https://github.com/Shopify/shopify_app
-const topLevelInteraction = (shop: string, prefix = '') => {
+const topLevelInteraction = (shop: string, host: string, prefix = '') => {
   return `(function() {
       function setUpTopLevelInteraction() {
         var TopLevelInteraction = new ITPHelper({
-          redirectUrl: "${prefix}/auth?shop=${encodeURIComponent(shop)}",
+          redirectUrl: "${prefix}/auth?shop=${encodeURIComponent(
+    shop,
+  )}&host=${host}",
         });
 
         TopLevelInteraction.execute();