+++ src/auth/create-request-storage-access.ts
@@ -1,9 +1,10 @@
 import {Context} from 'koa';
 
-import {OAuthStartOptions} from '../types';
-
 import Shopify from '@shopify/shopify-api';
+import fs from 'fs';
+import path from 'path';
 
+import {OAuthStartOptions} from '../types';
 import css from './client/polaris-css';
 import itpHelper from './client/itp-helper';
 import requestStorageAccess from './client/request-storage-access';
@@ -15,12 +16,16 @@ const BODY =
   'Your browser is blocking this app from accessing your data. To continue using this app, click Continue, then click Allow if the browser prompts you.';
 const ACTION = 'Continue';
 
+const APP_BRIDGE_SCRIPT = fs.readFileSync(
+  path.resolve(`${__dirname}/../app-bridge-2.0.12.js`),
+);
+
 export default function createRequestStorageAccess({
   prefix,
 }: OAuthStartOptions) {
   return function requestStorage(ctx: Context) {
     const {query} = ctx;
-    const {shop} = query;
+    const {shop, host} = query;
 
     if (shop == null) {
       ctx.throw(400, Error.ShopParamMissing);
@@ -38,12 +43,14 @@ export default function createRequestStorageAccess({
   <base target="_top">
   <title>Redirecting…</title>
 
+  <script>${APP_BRIDGE_SCRIPT}</script>
   <script>
     window.apiKey = "${Shopify.Context.API_KEY}";
+    window.host = "${host}";
     window.shopOrigin = "https://${encodeURIComponent(shop)}";
     ${itpHelper}
     ${storageAccessHelper}
-    ${requestStorageAccess(shop, prefix)}
+    ${requestStorageAccess(shop, host, prefix)}
   </script>
 </head>
 <body>